# docs/openapi-profile-api.yaml
openapi: 3.0.3
info:
  title: Profile API
  version: 1.0.0
  description: API para cadastro, consulta e atualização de perfis de assinatura digital.
  contact:
    name: Equipe de Desenvolvimento
    email: contato@empresa.com
    url: https://empresa.com
tags:
  - name: Profile
    description: Operações relacionadas a perfis de assinatura digital
servers:
  - url: http://localhost:8080
paths:
  /profiles:
    post:
      tags: [Profile]
      summary: Cria um novo perfil
      description: Cria um novo perfil de assinatura digital.
      operationId: createProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSignerProfileDTO'
      responses:
        '201':
          description: Perfil criado com sucesso
          headers:
            Location:
              description: URL do recurso criado
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
        '400':
          description: Requisição inválida
        '500':
          description: Erro interno
  /profiles/{id}:
    get:
      tags: [Profile]
      summary: Consulta um perfil por ID
      description: Retorna os dados de um perfil de assinatura digital pelo ID.
      operationId: getProfileById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Perfil encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignerProfile'
        '404':
          description: Perfil não encontrado
    patch:
      tags: [Profile]
      summary: Atualiza campos de um perfil
      description: Atualiza parcialmente os dados de um perfil de assinatura digital.
      operationId: updateProfile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSignerProfileDTO'
      responses:
        '200':
          description: Perfil atualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Requisição inválida
        '404':
          description: Perfil não encontrado
        '500':
          description: Erro interno
components:
  schemas:
    CreateSignerProfileDTO:
      type: object
      required:
        - signer
        - profile_id
      properties:
        signer:
          type: string
          description: Fabricante do dispositivo
          enum: ["positivo", "gertec"]
        profile_id:
          type: string
        description:
          type: string
        configs:
          type: array
          items:
            $ref: '#/components/schemas/ProfileConfig'
        upload:
          $ref: '#/components/schemas/TransferInfo'
        download:
          $ref: '#/components/schemas/TransferInfo'
    UpdateSignerProfileDTO:
      type: object
      properties:
        description:
          type: string
        configs:
          type: array
          items:
            $ref: '#/components/schemas/ProfileConfig'
        upload:
          $ref: '#/components/schemas/TransferInfo'
        download:
          $ref: '#/components/schemas/TransferInfo'
    SignerProfile:
      type: object
      properties:
        id:
          type: string
        signer:
          type: string
        profile_id:
          type: string
        description:
          type: string
        configs:
          type: array
          items:
            $ref: '#/components/schemas/ProfileConfig'
        upload:
          $ref: '#/components/schemas/TransferInfo'
        download:
          $ref: '#/components/schemas/TransferInfo'
    ProfileConfig:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
    TransferInfo:
      type: object
      properties:
        url:
          type: string
        tries:
          type: integer
        interval:
          type: integer